src/ingest/pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "Add description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python-ng>=2.2.2",
    "pyyaml>=6.0.3",
    "sseclient-py>=1.8.0",
    "requests>=2.31.0",
    "boto3>=1.34.0",
]
